+++ src/main/scala/org/idio/dbpedia/spotlight/stores/IdioContextStore.scala
@@ -1,9 +1,8 @@
-package org.idio.dbpedia.spotlight
+package org.idio.dbpedia.spotlight.stores
 
 import org.dbpedia.spotlight.db.memory.{MemoryStore, MemoryContextStore, MemoryTokenTypeStore}
 import java.io.{File, FileInputStream}
-import org.dbpedia.spotlight.model.{TokenType, Token, DBpediaResource}
-import scala.collection.mutable.HashMap
+import org.dbpedia.spotlight.model.DBpediaResource
 
 /**
  * Created by dav009 on 02/01/2014.
